Title: n개의 숫자 중 최소, Time: 158ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240530/n개의 숫자 중 최소/min-of-n-num.java
@@ -7,7 +7,7 @@ public class Main {
 
         int n = sc.nextInt();
         int[] arr = new int[n];
-        int min = (int)Math.pow(2,31) - 1;
+        double min = Math.pow(2,31) - 1;
         for(int i = 0 ; i < n ; i++){
             arr[i] = sc.nextInt();
             if(arr[i] <= min){
